refactor: adjust the structure of modular_alert_builder module (#682)
* refactor: move modular_alert_builder under commands folder

* refactor: remove some unused code from alerts module

* refactor: remove unnecessary build_core folder

* chore: adjust license information
