+++ example/globalConfig.json
@@ -161,7 +161,7 @@
     "meta": {
         "name": "Splunk_TA_dummy_data",
         "restRoot": "Splunk_TA_dummy_data",
-        "version": "5.21.0R18105c14",
+        "version": "5.21.0R43e797b8",
         "displayName": "Splunk_TA_dummy_data",
         "schemaVersion": "0.0.3"
     }
