+++ splunk_add_on_ucc_framework/commands/modular_alert_builder/__init__.py
@@ -17,13 +17,13 @@ import logging
 import os
 import shutil
 
-from splunk_add_on_ucc_framework.modular_alert_builder.build_core.alert_actions_conf_gen import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_conf_gen import (
     generate_alert_actions_conf,
 )
-from splunk_add_on_ucc_framework.modular_alert_builder.build_core.alert_actions_html_gen import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_html_gen import (
     generate_alert_actions_html_files,
 )
-from splunk_add_on_ucc_framework.modular_alert_builder.build_core.alert_actions_py_gen import (
+from splunk_add_on_ucc_framework.commands.modular_alert_builder.alert_actions_py_gen import (
     generate_alert_actions_py_files,
 )
