+++ tests/testdata/test_addons/package_global_config_inputs_configuration_alerts/globalConfig.json
@@ -1108,7 +1108,7 @@
     "meta": {
         "name": "Splunk_TA_UCCExample",
         "restRoot": "splunk_ta_uccexample",
-        "version": "5.21.0R18105c14",
+        "version": "5.21.0R43e797b8",
         "displayName": "Splunk UCC test Add-on",
         "schemaVersion": "0.0.3"
     }
